Remove Kaputt and add Confidence
Since you published your very detailed comparison I worked intensively
on Kaputt and republished that as Confidence — a more optimistic name
that underlines it's a tool to deploy code in confidence.

This PR is my best effort to update your detailed comparison in
consequence of the name change and the feature change.  Since then,
Confidence is used in other projects, such as KONS-9.

While I continue to develop it, I feel it was a good time to try to
update your article, so that readers will find information about
Confidence there.

- I tried to update the text according to the feature changes.
- I renamed and reordered lists.
- I did my best to format a list of assertions on two columns.
- I did NOT regenerate HTML version of the document.
- I did NOT rerun the time measurement you did.

However I am interested adding these time measurements as a benchmark
to Confidence and add your test framework testsuite to Confidence tests.